requirements.txt
alembic==1.14.1
aniso8601==10.0.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.14.3
blinker==1.8.2
cachelib==0.13.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
cloudinary==1.36.0
decorator==5.2.1
Deprecated==1.3.1
exceptiongroup==1.3.0
executing==2.2.1
Faker==35.2.2
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==1.2.1
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==21.2.0
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
ipdb==0.13.9
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
limits==3.13.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
matplotlib-inline==0.1.7
mdurl==0.1.2
MouseInfo==0.1.3
ordered-set==4.1.0
packaging==24.2
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
pluggy==1.5.0
prometheus_client==0.21.1
prometheus_flask_exporter==0.23.2
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.9.0
PyMsgBox==2.0.1
pyngrok==7.2.12
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
PyYAML==6.0.2
redis==5.0.0
referencing==0.35.1
requests==2.31.0
rich==13.9.4
rpds-py==0.20.1
sendgrid==6.11.0
sentry-sdk==2.48.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.36
SQLAlchemy-serializer==1.4.12
stack-data==0.6.3
starkbank-ecdsa==2.2.0
toml==0.10.2
tomli==2.3.0
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==2.2.3
wcwidth==0.2.14
Werkzeug==3.0.6
wikipedia==1.4.0
wrapt==2.0.1
WTForms==3.1.2
zipp==3.20.2
